Update factsheet.md
Could you adapt the side width? The slider for the horizontal axis is at the bottom of the page and thus not handy to see how a sentence continues (e.g. line 39)
Line 13: Some of the special populations mentioned need to be incorporated manually (HI, RI, Obese)
Why do we have a dedicated section on "Permeabilits" (line 114) although this is already covered in the compound section?
blabla (line 257)?
Features of MoBi could be listed in the same tabular style like for PK-Sim